+++ src/main/java/META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-duke.Main-Class: duke.Duke
+duke.gui.Main-Class: duke.Duke
