+++ src/main/java/duke/command/ByeCommand.java
@@ -1,9 +1,10 @@
 package duke.command;
 
-import duke.DukeException;
-import duke.Storage;
-import duke.TaskList;
 import duke.ui.Ui;
+import duke.util.DukeException;
+import duke.util.Storage;
+import duke.util.TaskList;
+
 
 /**
  * Represents the command to exit the program.
