+++ src/main/java/duke/command/DeadlineCommand.java
@@ -1,10 +1,11 @@
 package duke.command;
 
-import duke.DukeException;
-import duke.Storage;
-import duke.TaskList;
 import duke.task.Deadline;
 import duke.ui.Ui;
+import duke.util.DukeException;
+import duke.util.Storage;
+import duke.util.TaskList;
+
 
 
 /**
