+++ src/main/java/duke/util/DukeException.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.util;
 
 public class DukeException extends Exception {
     public DukeException(String errorMessage) {
